fix: 버스트 시 승패 오류 수정
